requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.2
django-admin-interface==0.29.1
django-colorfield==0.11.0
django-crispy-forms==2.3
django-registration-redux==2.13
django-versatileimagefield==3.1
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-magic==0.4.27
python-slugify==8.0.4
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
